OE-176 fix edit page when change change financer to enium. and calc margin

+++ app/Http/Livewire/Customer/Create.php
@@ -182,7 +182,7 @@ class Create extends Component
     public function calculateGrossRepComission(Customer $customer)
     {
         if ( $customer->margin >= 0 && $customer->system_size >= 0 ) {
-            return intval($customer->margin) * intval($customer->system_size) * 1000;
+            return floatval($customer->margin) * floatval($customer->system_size) * 1000;
         } else {
             return 0;
         }
